test(ui): split stored-snapshot hydration coverage into its own file
chat-pane.test.ts had grown past the lint line budget. Move the
persistent-snapshot hydration tests (including the new ordering-fence
regression) into chat-pane-snapshot-hydration.test.ts and share the
pane initialization fixtures through the existing test support module.
