+++ ui/src/pages/chat/chat-pane.test.ts
@@ -1,11 +1,8 @@
 /* @vitest-environment jsdom */
 
-import { IDBFactory } from "fake-indexeddb";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { createDeferred } from "../../../../test/helpers/promise.js";
-import type { GatewayBrowserClient } from "../../api/gateway.ts";
 import type { GatewaySessionRow } from "../../api/types.ts";
-import { createChatAttachmentHandoff } from "../../app/chat-attachment-handoff.ts";
 import type { ApplicationContext } from "../../app/context.ts";
 import { createInitialUserMessageHandoff } from "../../app/initial-user-message-handoff.ts";
 import { t } from "../../i18n/index.ts";
@@ -18,21 +15,16 @@ import { loadChatHistory } from "./chat-history.ts";
 import { subscribeChatPaneSnapshotInvalidation } from "./chat-pane-startup-subscriptions.ts";
 import {
   createGatewayBrowserClientFixture,
+  createInitializationContext,
   createSessionCapabilityFixture,
   createSessionContext,
   createTestChatPane,
+  nativeHistoryMessage,
   type TestChatPane,
 } from "./chat-pane.test-support.ts";
 import type { ChatPageHost } from "./chat-state-host.ts";
 import type { SidebarContent } from "./components/chat-sidebar.ts";
-import { prepareInitialUserMessageHandoff } from "./initial-turn-handoff.ts";
-import {
-  cacheChatSessionSnapshot,
-  observeChatCache,
-  type ChatMessageCache,
-} from "./session-message-cache.ts";
-import { clearStoredChatSnapshots } from "./session-snapshot-invalidation.ts";
-import { SessionSnapshotStore } from "./session-snapshot-store.ts";
+import { cacheChatSessionSnapshot, type ChatMessageCache } from "./session-message-cache.ts";
 import { openSlot } from "./sidebar-layout.ts";
 
 vi.mock("../../lib/toast.ts", () => ({ showToast: vi.fn() }));
@@ -51,67 +43,6 @@ function dispatchSidebarShortcut(pane: TestChatPane, shiftKey = true) {
   });
   pane.handleDocumentKeydown(event);
   return event;
-}
-
-function createInitializationContext(): ApplicationContext {
-  return {
-    basePath: "",
-    gateway: {
-      snapshot: {
-        client: null,
-        phase: "stopped",
-        offlineStable: false,
-        hello: null,
-        canvasPluginSurfaceUrl: null,
-        assistantAgentId: null,
-        sessionKey: "",
-        lastError: null,
-        lastErrorCode: null,
-      },
-      subscribe: () => () => {},
-      subscribeEvents: () => () => {},
-    },
-    config: {
-      current: {
-        assistantIdentity: {
-          agentId: null,
-          name: "Assistant",
-          avatar: null,
-          avatarSource: null,
-          avatarStatus: null,
-          avatarReason: null,
-        },
-        serverVersion: null,
-        localMediaPreviewRoots: [],
-        embedSandboxMode: "strict",
-        allowExternalEmbedUrls: false,
-        terminalEnabled: false,
-      },
-    },
-    agentSelection: { state: { selectedId: "main" } },
-    agents: { state: { agentsList: null } },
-    runtimeConfig: {
-      state: { configNeedsApply: false, configSnapshot: null },
-      subscribe: () => () => {},
-    },
-    placementStartup: {
-      get: () => null,
-      retry: () => undefined,
-      subscribe: () => () => {},
-    },
-    navigate: () => undefined,
-    initialUserMessage: createInitialUserMessageHandoff(),
-    chatAttachmentHandoff: createChatAttachmentHandoff(),
-    sessions: { state: { modelOverrides: {} } },
-  } as unknown as ApplicationContext;
-}
-
-function nativeHistoryMessage(seq: number, text = `message ${seq}`) {
-  return {
-    role: seq % 2 === 0 ? "assistant" : "user",
-    content: [{ type: "text", text }],
-    __openclaw: { seq },
-  };
 }
 
 describe("chat pane header state", () => {
@@ -565,176 +496,6 @@ describe("chat pane initialization", () => {
         totalMessages: 2,
       });
       expect(attachedState?.currentSessionId).toBe("split-session");
-    } finally {
-      pane.disconnectedCallback();
-    }
-  });
-
-  it("paints a persistent snapshot while the network refresh is already in flight", async () => {
-    vi.stubGlobal("indexedDB", new IDBFactory());
-    const targetSessionKey = "agent:main:persistent";
-    const cachedMessages = [nativeHistoryMessage(1, "persistent history")];
-    const networkMessages = [nativeHistoryMessage(1, "network history")];
-    const writer = new SessionSnapshotStore();
-    writer.write(targetSessionKey, {
-      messages: cachedMessages,
-      pagination: { hasMore: false, completeSnapshot: true },
-      sessionId: "persistent-session",
-    });
-    await writer.flush();
-    const response = createDeferred<Record<string, unknown>>();
-    const request = vi.fn(() => response.promise);
-    const pane = document.createElement("openclaw-chat-pane") as unknown as TestChatPane;
-    vi.spyOn(pane, "requestUpdate").mockImplementation(() => undefined);
-    vi.spyOn(pane, "performUpdate").mockImplementation(() => undefined);
-    const sharedMessages: ChatMessageCache = new Map();
-    const store = new SessionSnapshotStore(sharedMessages);
-    store.connect();
-    observeChatCache(sharedMessages, store);
-    pane.sessionKey = targetSessionKey;
-    pane.chatMessagesBySession = sharedMessages;
-    pane.sessionSnapshotStore = store;
-    pane.context = createInitializationContext();
-    const client = { request } as unknown as GatewayBrowserClient;
-    const stopAfterAttach = new Error("stop after attach");
-    let attachedState: ChatPageHost | undefined;
-    vi.spyOn(pane.chatState, "attach").mockImplementation((state) => {
-      attachedState = state;
-      state.client = client;
-      state.connected = true;
-      state.connectionEpoch = 1;
-      void loadChatHistory(state);
-      throw stopAfterAttach;
-    });
-
-    try {
-      expect(() => pane.connectedCallback()).toThrow(stopAfterAttach);
-      expect(request).toHaveBeenCalledWith(
-        "chat.history",
-        expect.objectContaining({ sessionKey: targetSessionKey }),
-      );
-      await vi.waitFor(() => expect(attachedState?.chatMessages).toEqual(cachedMessages));
-
-      response.resolve({ messages: networkMessages, sessionId: "network-session" });
-      await vi.waitFor(() => expect(attachedState?.chatMessages).toEqual(networkMessages));
-    } finally {
-      pane.disconnectedCallback();
-      store.disconnect();
-      await store.whenIdle();
-      await clearStoredChatSnapshots();
-    }
-  });
-
-  it("discards persistent hydration when the network snapshot lands first", async () => {
-    vi.stubGlobal("indexedDB", new IDBFactory());
-    const targetSessionKey = "agent:main:network-first";
-    const writer = new SessionSnapshotStore();
-    writer.write(targetSessionKey, {
-      messages: [nativeHistoryMessage(1, "stale persistent history")],
-      pagination: { hasMore: false, completeSnapshot: true },
-      sessionId: "persistent-session",
-    });
-    await writer.flush();
-    const networkMessages = [nativeHistoryMessage(1, "authoritative network history")];
-    const request = vi.fn(async () => ({
-      messages: networkMessages,
-      sessionId: "network-session",
-    }));
-    const pane = document.createElement("openclaw-chat-pane") as unknown as TestChatPane;
-    vi.spyOn(pane, "requestUpdate").mockImplementation(() => undefined);
-    vi.spyOn(pane, "performUpdate").mockImplementation(() => undefined);
-    const sharedMessages: ChatMessageCache = new Map();
-    const store = new SessionSnapshotStore(sharedMessages);
-    store.connect();
-    observeChatCache(sharedMessages, store);
-    pane.sessionKey = targetSessionKey;
-    pane.chatMessagesBySession = sharedMessages;
-    pane.sessionSnapshotStore = store;
-    pane.context = createInitializationContext();
-    const client = { request } as unknown as GatewayBrowserClient;
-    const stopAfterAttach = new Error("stop after attach");
-    let attachedState: ChatPageHost | undefined;
-    vi.spyOn(pane.chatState, "attach").mockImplementation((state) => {
-      attachedState = state;
-      state.client = client;
-      state.connected = true;
-      state.connectionEpoch = 1;
-      void loadChatHistory(state);
-      throw stopAfterAttach;
-    });
-
-    try {
-      expect(() => pane.connectedCallback()).toThrow(stopAfterAttach);
-      await vi.waitFor(() => expect(attachedState?.chatMessages).toEqual(networkMessages));
-      await new Promise<void>((resolve) => {
-        setTimeout(resolve, 0);
-      });
-      expect(attachedState?.chatMessages).toEqual(networkMessages);
-    } finally {
-      pane.disconnectedCallback();
-      store.disconnect();
-      await store.whenIdle();
-      await clearStoredChatSnapshots();
-    }
-  });
-
-  it("keeps an admitted first-turn prompt visible when stored hydration resolves late", async () => {
-    const targetSessionKey = "agent:main:first-turn-retry";
-    const client = {
-      addEventListener: vi.fn(() => vi.fn()),
-      request: vi.fn(),
-    } as unknown as GatewayBrowserClient;
-    const context = createInitializationContext();
-    context.gateway.snapshot.client = client;
-    const pane = document.createElement("openclaw-chat-pane") as unknown as TestChatPane;
-    vi.spyOn(pane, "requestUpdate").mockImplementation(() => undefined);
-    vi.spyOn(pane, "performUpdate").mockImplementation(() => undefined);
-    pane.sessionKey = targetSessionKey;
-    pane.chatMessagesBySession = new Map();
-    let deliverStoredSnapshot: ((snapshot: unknown) => void) | undefined;
-    pane.sessionSnapshotStore = {
-      read: () =>
-        new Promise((resolve) => {
-          deliverStoredSnapshot = resolve;
-        }),
-    } as never;
-    pane.context = context;
-    prepareInitialUserMessageHandoff(
-      context.initialUserMessage,
-      targetSessionKey,
-      { attachments: [], createdAt: 1, text: "retry the rejected prompt" },
-      client,
-      { runId: "initial-run" },
-    );
-    const stopAfterAttach = new Error("stop after attach");
-    let attachedState: ChatPageHost | undefined;
-    vi.spyOn(pane.chatState, "attach").mockImplementation((state) => {
-      attachedState = state;
-      throw stopAfterAttach;
-    });
-
-    try {
-      expect(() => pane.connectedCallback()).toThrow(stopAfterAttach);
-      expect(attachedState?.chatMessages).toEqual([
-        expect.objectContaining({
-          role: "user",
-          content: [expect.objectContaining({ type: "text", text: "retry the rejected prompt" })],
-        }),
-      ]);
-      deliverStoredSnapshot?.({
-        messages: [nativeHistoryMessage(1, "stale stored transcript")],
-        pagination: { hasMore: false, completeSnapshot: true },
-        sessionId: "stored-session",
-      });
-      await Promise.resolve();
-      await Promise.resolve();
-      expect(deliverStoredSnapshot).toBeDefined();
-      expect(attachedState?.chatMessages).toEqual([
-        expect.objectContaining({
-          role: "user",
-          content: [expect.objectContaining({ type: "text", text: "retry the rejected prompt" })],
-        }),
-      ]);
     } finally {
       pane.disconnectedCallback();
     }
